Allow IPv6 node addresses for geneve tunnels
We didn't hit this on AWS or Azure because both of those platforms
have IPv4 and IPv6 addresses for nodes, so this would succeed using
the IPv4 address.  On a single-stack bare metal env, we must allow
IPv6 addresses here.

The result of this code is the IP used to create geneve tunnels
between Nodes.