Fixes accidental override of file conf settings with defaults
When config is read, the config file is parsed first. After these
settings are loaded, the CLI settings are parsed and override any
settings read from the file. This is fine, except for the CLI settings
may be defaulted, which will just override any specified config file
settings with defaults.

This patch addresses the problem by checking during config override if
the the configuration option has a default specified, and if so, ignores
overriding the current config option.

Additionally, this patch adds real default values for settings.

Signed-off-by: Tim Rozet <trozet@redhat.com>